+++ ImP/preProcessing/preProcessor.py
@@ -2,70 +2,57 @@ import cv2
 import numpy as np
 from skimage import data
 from skimage import exposure 
-from PIL import Image, ImageFilter
+from PIL import Image
+
 
 class PreProcessor:
+    """
+    TODO: Perhaps make a new task to develop a "main" function that takes the newly taken image and takes it through a
+     process that first runs image similarity, then the low/high contrast check, and finally adjust the image until it passes the low/high
+     contrast check.
+    """
+    image_similarity_threshold = 2.2
 
-	threshold_value = 2.2
+    def __init__(self, pathToNewImage):
+        self.image = cv2.imread(pathToNewImage) # for OpenCV
+        self.scikitImage = data.imread(pathToNewImage) # for scikit
+        self.pilImage = Image.open(pathToNewImage) # for PIL
+        self.path = pathToNewImage # used to save modified images to the same location, overriding them essentially
 
-	def __init__(self, pathToNewImage):
-		self.image = cv2.imread(pathToNewImage) #for OpenCV
-		self.scikitImage = data.imread(pathToNewImage) # for scikit 
-		self.pilImage = Image.open(pathToNewImage) # for PIL 
-		#only using this var underneath to play with new images at the moment
-		self.path = pathToNewImage
+    def is_similar(self, pathToOtherImage):
+        existingImage = cv2.imread(pathToOtherImage)
+        difference = cv2.subtract(self.image, existingImage)
+        self.pilImage.close()
+        return np.mean(difference) <= self.image_similarity_threshold
 
-	def is_similar(self, pathToOtherImage):
-		existingImage = cv2.imread(pathToOtherImage)
-		difference = cv2.subtract(self.image, existingImage)
-		self.pilImage.close()
-		return np.mean(difference) <= self.threshold_value
-		
+    def darken_image(self, darkFactor):
+        """
+        By setting the darkfactor to anything less than 1.0 we can darken the image.
+        """
+        self.pilImage.point(lambda x: x*darkFactor).save(self.path)
 
-	def increase_contrast(self):
-		pass
+    def brighten_image(self, brightFactor):
+        """
+        Works the same way as darken_image() but x>1.0 to brighten the image as opposed to
+        being x<1.0 to darken it.
+        """
+        self.pilImage.point(lambda x: x*brightFactor).save(self.path)
 
+    def is_low_contrast(self, threshold):
+        """
+        By default the threshold value is set to 0.05. We override that amount by setting our own threshold.
 
-	def darken_image(self, darkFactor):
-		"""
-		Aruco knows no bounds when it comes to dark images. At the moment we do not have a
-		set value for how dark we can make an image. It works under x =0.2, but that may
-		not be the case under our algorithm (pycuda version).
-		
-		TODO: Find threshold value for both pycuda and regular algorithm
-			  Also, modify so we can set darkness levels via arg		
-		"""
-		#image = Image.open(self.path)
-		#image.point(lambda x: x*0.4).save('darkest.jpg')
-		self.pilImage.point(lambda x: x*darkFactor).save(self.path)
-		#self.pilImage = self.pilImage.point(lambda x: x*0.6)
-		#self.pilImage.close()
+        TODO: Find threshold we will use.
+        """
+        return exposure.is_low_contrast(self.scikitImage, fraction_threshold=threshold)
 
-	def brighten_image(self, brightFactor):
-		"""
-		Works the same way as darken_image() but x>1.0 to brighten the image as ooposed to 
-		being x<1.0 to darken it.
+    def is_too_bright(self, threshold):
+        """
+        Here we simply use a higher threshold to inverse the use of the exposure.is_low_contrast and detect
+        high contrast.
 
-		TODO: Develop way to find out how much we need to brighten the image
-			  Feed in x value via arg
-
-		"""
-		
-		self.pilImage.point(lambda x: x*brightFactor).save(self.path)
-		#self.pilImage.close()
-
-	def restore_image(self, factor):
-		self.pilImage.point(lambda x: x/factor).save(self.path)
-
-	def is_low_contrast(self, threshold):
-		"""
-		By default the threshold value is set to 0.05, however under darker_image set to 
-		0.4 we achieve a low contrast==True under the threshold of 0.24
-		"""
-		return exposure.is_low_contrast(self.scikitImage, fraction_threshold=threshold)
-		
-	
-	def is_too_bright(self, threshold):
-		return exposure.is_low_contrast(self.scikitImage, fraction_threshold=threshold)
+        TODO: Find threshold we will use.
+        """
+        return exposure.is_low_contrast(self.scikitImage, fraction_threshold=threshold)
